Update Firefox Lite functional tests (Fixes #9119)
