+++ tests/pages/firefox/whatsnew/whatsnew_africa.py
@@ -9,7 +9,7 @@ from pages.base import BasePage
 
 class FirefoxWhatsNewAfricaPage(BasePage):
 
-    URL_TEMPLATE = '/{locale}/firefox/whatsnew/africa/'
+    URL_TEMPLATE = '/{locale}/firefox/79.0/whatsnew/africa/'
 
     _qr_code_locator = (By.CSS_SELECTOR, '.lite-qrcode-container > img')
